signup singin

+++ controllers/game.controller.js
@@ -1,6 +1,7 @@
 const CustomError = require("../errors");
 const Game = require("../model/Game");
-const { uploadImage } = require("../utils/cloudinary");
+const { uploadImage, deleteImage } = require("../utils/cloudinary");
+const getInfo = require("../utils/info");
 const fs = require("fs-extra");
 
 const getGame = async (req, res, next) => {
@@ -21,8 +22,14 @@ const getGame = async (req, res, next) => {
 
 const getAllGames = async (req, res, next) => {
   try {
-    const games = await Game.find();
-    res.status(200).json({ success: true, games });
+    const count = await Game.count();
+    const info = getInfo(count, req);
+
+    const games = await Game.find()
+      .limit(info.limit)
+      .skip((info.page - 1) * info.limit);
+
+    res.status(200).json({ success: true, info, games });
   } catch (err) {
     next(err);
   }
@@ -33,26 +40,36 @@ const createGame = async (req, res, next) => {
     throw new CustomError.BadRequestError("No thumbail upload");
   }
 
+  if (!req.files?.screenshots) {
+    throw new CustomError.BadRequestError("No screenshots upload");
+  }
+
   try {
-    const thumbnail = await uploadImage(req.files.thumbnail.tempFilePath);
+    const header = await uploadImage(req.files.thumbnail.tempFilePath);
+
+    const screenshots = [];
+    for (let i = 0; i < req.files.screenshots.length; i++) {
+      const item = req.files.screenshots[i];
+      let img = await uploadImage(item.tempFilePath);
+      await fs.unlink(item.tempFilePath);
+      screenshots.push({
+        screenshot_url: img.secure_url,
+        screenshot_id: img.public_id,
+      });
+    }
+
+    const requirements = {
+      os: req.body.os,
+      processor: req.body.processor,
+      graphics: req.body.graphics,
+      storage: req.body.storage,
+    };
 
     const newGame = new Game({
-      title: req.body.title,
-      thumbnail: thumbnail.secure_url,
-      description: req.body.description,
-      genre: req.body.genre,
-      publisher: req.body.publisher,
-      release_date: req.body.release_date,
-      platform: req.body.platform,
-      franchise: req.body.franchise,
-      developer: req.body.developer,
-      minimum_system_requirements: {
-        os: req.body.os,
-        processor: req.body.processor,
-        graphics: req.body.graphics,
-        storage: req.body.storage,
-      },
-      screenshots: req.body.screenshots,
+      ...req.body,
+      thumbnail: { url: header.secure_url, img_id: header.public_id },
+      screenshots,
+      minimum_system_requirements: requirements,
     });
 
     await fs.unlink(req.files.thumbnail.tempFilePath);
@@ -66,9 +83,21 @@ const createGame = async (req, res, next) => {
 
 const updateGame = async (req, res, next) => {
   const id = req.params.id;
+  const update = req.body;
 
   try {
-    const gameUpdate = await Game.findByIdAndUpdate(id, req.body, {
+    if (req.files?.thumbnail) {
+      const thumbnail = await uploadImage(req.files.thumbnail.tempFilePath);
+      update.thumbnail = {
+        url: thumbnail.secure_url,
+        img_id: thumbnail.public_id,
+      };
+      const game = await Game.findById(id);
+      await deleteImage(game.thumbnail.img_id);
+      await fs.unlink(req.files.thumbnail.tempFilePath);
+    }
+
+    const gameUpdate = await Game.findByIdAndUpdate(id, update, {
       new: true,
     });
 
@@ -90,6 +119,13 @@ const deleteGame = async (req, res, next) => {
 
     if (!deletedGame) {
       throw new CustomError.NotFoundError(`No game with id : ${id}`);
+    }
+
+    await deleteImage(deletedGame.thumbnail.img_id);
+
+    for (let i = 0; i < deletedGame.screenshots.length; i++) {
+      const item = deletedGame.screenshots[i];
+      await deleteImage(item.screenshot_id);
     }
 
     await deletedGame.remove();
